Convert arenas
